refactor: drop obsolete model stub
