--- app/models/user.py
@@ -1,5 +1,0 @@
-# Stub for app.models.user to satisfy mypy and test imports 
-
-# Stub for User model to satisfy import errors in tests
-class User:
-    pass
